writer_ctrl: make fifo and wb interfaces non-registered
The extra latency by the registering of the fifo signals
result in that the fifo count information does not have time
to propagate back when it is used in S_IDLE state.
The extra latency on the wb signals caused cyc&stb to assert
even when the slave deasserts ack.